Add Tiberian Sun Map Generator
Implements a purpose-built map generator for Tiberian Sun.

This makes use of many of the same underlying utilities as the
existing map generators, but also adds a few more utilities:

- RampTiler: for placing ramp tiles to best-fit a target heightmap.
- LatTiler: for smoothing out terrain transistions using LAT tiles.
- Support for mods where resource densities are baked into a map.
- Support for zoning based on ramps.

In order to support ramp tiling, additional "Riser" information is
added to the TS tileset YAML to specify height discontinuities between
tiles.

Key Features:

- Support for both Temperate and Snow tilesets.
- Largely similar options to the generator used by CnC.
- Variable terrain height with smooth transistions between cliffs.
- Configurable flat, cliffy, or mixed coastlines.
- Mixed patches of ground tiles.
- Optional veinholes.

The map generator primarily aims to provide a reasonable example of a
RectangularIsometric map generator. It does not implement the full set
of features available to TS maps. For example, it doesn't make use of
all environments (urban areas, tiberium corruption); doesn't support
roads, bridges, or tunnels; and does not add customized lighting.

The map generator can be used via either the lobby or the map editor.